requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
langchain_community
pypdf
bs4
langchain-text-splitters
langchain-openai
langchain
faiss-cpu
chromadb
langchain_chroma
langsmith
langchain_ollama
beautifulsoup4
streamlit
langchain_groq
groq
uvicorn
sse_starlette
pydantic==2.9.2 
fastapi==0.115.0 
langserve==0.3.0